chapter_02_cryptography/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
fastapi==0.116.1
idna==3.10
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
